+++ experiments/debug_no_checkpointing.py
@@ -1,0 +1,9 @@
+import ray
+
+from parameter_servers.server_actor import ParameterServer
+
+LEARNING_RATE = 1e-2
+
+def run_debug_no_checkpointing(model):
+  ps = ParameterServer.remote(LEARNING_RATE)
+  ray.get([ps.run_training.remote(False)])
